BAYL-48 получение информацию о проекте

+++ src/main/java/org/baylist/todoist/api/Todoist.java
@@ -8,6 +8,6 @@ public interface Todoist {
 
     List<Project> getProjects();
 
-    Project getProject(int index);
+    Project getProject(long index);
 
 }
